refac: Line에 있는 getOrderedStations 함수를 Stations로 이동 및 객체 비교시 equals로 비교
